+++ lib/src/elements/th_bezier_curve_line_segment.dart
@@ -25,6 +25,8 @@ class THBezierCurveLineSegment extends THLineSegment with THHasOptionsMixin {
     required this.controlPoint1,
     required this.controlPoint2,
     required super.endPoint,
+    super.optionsMap,
+    super.attrOptionsMap,
     super.originalLineInTH2File = '',
   }) : super.withEndPoint();
 
@@ -34,6 +36,8 @@ class THBezierCurveLineSegment extends THLineSegment with THHasOptionsMixin {
     required List<dynamic> controlPoint1,
     required List<dynamic> controlPoint2,
     required List<dynamic> endPoint,
+    super.optionsMap,
+    super.attrOptionsMap,
     super.originalLineInTH2File = '',
   }) : super.withoutEndPoint() {
     this.controlPoint1 = THPositionPart.fromStringList(list: controlPoint1);